Start modifying towards using instead stockstats library
