+++ src/utils.py
@@ -4,25 +4,32 @@
 utils.py
 Module containing util general functions
 """
+def bought_status(rsi):
+    """
+    Check if the rsi is overbought or oversold
+
+    @arg rsi. Number containing the value of the rsi
+    @return is_upwards, is_downwards
+    """
+    return rsi >= 66, rsi <=33
+
+def macd_trend(macd, signal):
+    """
+    Checks if the current trend is is_upwards or downards
+
+    @arg macd. List containing the macd values
+    @arg signal. List containing the signal
+    @return is_upwards, is_downward
+    """
+    is_upwards = signal[-2] > macd[-2] and signal[-1] <= macd[-1]
+    is_downward = signal[-2] < macd[-2] and signal[-1] >= macd[-1]
+    return is_upwards, is_downward
 
 def divide(numerator, denominator):
     """
     Divide function to avoid errors by dividing over zero
     """
     return 0 if denominator == 0 else numerator / denominator
-
-def calculate_predictions(true_positive, false_positive, true_negative, false_negative):
-    """
-    Calculate predictions based on the number of true/false positive/negative's
-    """
-    sens_denom = true_positive + false_negative
-    spec_denom = true_negative + false_positive
-    sensitivity = divide(true_positive, sens_denom)
-    specificity = divide(true_negative, spec_denom)
-    accuracy = divide(true_positive + true_negative, sens_denom + spec_denom)
-    tpr = sensitivity  # Calculate the true positive rate
-    fpr = 1 - specificity  # Calculate the false positive rate
-    return accuracy, tpr, fpr
 
 def chunks(lst, amount):
     """
